Implement the beginnings of Stores
Stores are special case Action objects that have a list of items they
sell. When the engine comes upon a store, the event loop blocks until
the user leaves the store while recording the items that the user wants
to purchase.

There is currently no validation for anything here.
The following can be considered TODOs
* Money isn't checked or deducted
* Stores can't yet specify the quantity of a given item
* Items aren't yet added to the player's inventory