Implement the beginnings of issues
There are two issue types: health and mechanical. Issues happen randomly
(not iplemented yet) based on stats and can be dependent on a player
having certain items in their inventory or the player having performed
certain actions in the past.

At this point, the library is feature complete for a minimilly viable
product. There are so many rules that need to be handled still, the
client needs to be finished, and the story needs to be written. However,
before all of that, I think its time for a nice refactor....