Huge refactor
* Cleaned up foward declarations
* Models and their managing objects are in separate classes
* Game objects manage all other objets
* All object getters are by reference now. This makes passing things w/
std::future easy. Each object provides a constant empty version of
itself for when the getter fails. == is overloaded to handle checking
for the empty case (works for all cases)
* Tests are in their own files
* Responsibilities of the engine are better separated
* Many small bug fixes

Some tests were removed to make this all work and will be readded (with
more test cases) in a future patch.